requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn<1.6
snfpy
scikit-survival
ipython
torch
umap-learn
hdbscan
ipykernel
jupyter
ipywidgets
requests